Done with the profile editing. Fixed the order total error.

+++ Jio/Models/Menu_Item.cs
@@ -12,7 +12,7 @@ namespace Jio.Models
         public int RestaurantID { get; set; }
         public string Item_Name { get; set; }
         public string description { get; set; }
-        public int price { get; set; }
+        public decimal price { get; set; }
         public Restaurant restaurant { get; set; }
         //public virtual ICollection<Restaurant> restaurant { get; set; }
